# DictionarySlim backport improvements, retaining more entropy