contributes to https://github.com/dotnet/corefx/issues/33392

Use uint hashcode to retain more entropy

before 

``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=Windows 10.0.17134.590 (1803/April2018Update/Redstone4)
Intel Core i7 CPU 860 2.80GHz (Nehalem), 1 CPU, 8 logical and 4 physical cores
Frequency=2727535 Hz, Resolution=366.6314 ns, Timer=TSC
.NET Core SDK=3.0.100-preview-010184
  [Host]     : .NET Core 3.0.0-preview-27324-5 (CoreCLR 4.6.27322.0, CoreFX 4.7.19.7311), 64bit RyuJIT
  Job-SOCKON : .NET Core f01bf9f8-609f-443f-b1f9-73654bd6fdd8 (CoreCLR 4.6.27523.0, CoreFX 4.7.19.12501), 64bit RyuJIT
  Job-LPCBGO : .NET Core f01bf9f8-609f-443f-b1f9-73654bd6fdd8 (CoreCLR 4.6.27523.0, CoreFX 4.7.19.12501), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                     Namespace |                             Type |        Method | InvocationCount | UnrollFactor | Size |   Items | Count |                 Mean |             Error |            StdDev |               Median |                  Min |                  Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------ |--------------------------------- |-------------- |---------------- |------------- |----- |-------- |------ |---------------------:|------------------:|------------------:|---------------------:|---------------------:|---------------------:|------------:|------------:|------------:|--------------------:|
|            **System.Collections** |           **CtorDefaultSize&lt;Int32&gt;** |    **Dictionary** |               **1** |           **16** |    **?** |       **?** |     **?** |            **21.138 ns** |         **0.5386 ns** |         **0.4498 ns** |            **21.034 ns** |            **20.602 ns** |            **22.054 ns** |      **0.0171** |           **-** |           **-** |                **72 B** |
|            System.Collections |          CtorDefaultSize&lt;String&gt; |    Dictionary |               1 |           16 |    ? |       ? |     ? |           115.775 ns |         2.1777 ns |         2.0370 ns |           115.759 ns |           112.713 ns |           119.546 ns |      0.0169 |           - |           - |                72 B |
| **System.Collections.Concurrent** |                   **IsEmpty&lt;Int32&gt;** |    **Dictionary** |               **1** |           **16** |    **0** |       **?** |     **?** |           **167.668 ns** |         **1.7925 ns** |         **1.6767 ns** |           **167.952 ns** |           **164.903 ns** |           **171.290 ns** |           **-** |           **-** |           **-** |                   **-** |
| System.Collections.Concurrent |                  IsEmpty&lt;String&gt; |    Dictionary |               1 |           16 |    0 |       ? |     ? |           175.245 ns |         3.1680 ns |         2.9634 ns |           175.101 ns |           170.252 ns |           181.164 ns |           - |           - |           - |                   - |
|            **System.Collections** |        **DictionaryMappingFunction** |       **Entropy** |               **1** |           **16** |    **?** |     **500** |     **?** |        **85,153.309 ns** |       **967.4673 ns** |       **904.9695 ns** |        **85,026.696 ns** |        **84,090.283 ns** |        **87,115.608 ns** |     **17.1371** |      **0.3360** |           **-** |             **73160 B** |
|            **System.Collections** |         **TryAddDefaultSize&lt;Int32&gt;** |    **Dictionary** |               **1** |           **16** |    **?** |       **?** |  **2048** |       **151,853.548 ns** |     **1,614.4616 ns** |     **1,510.1684 ns** |       **152,020.658 ns** |       **149,280.358 ns** |       **154,742.987 ns** |     **24.5098** |     **12.2549** |           **-** |            **154192 B** |
|            System.Collections |        TryAddDefaultSize&lt;String&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |       260,378.396 ns |     3,333.4320 ns |     3,118.0943 ns |       260,265.627 ns |       254,684.640 ns |       266,279.359 ns |     46.1066 |     40.9836 |     26.6393 |            215717 B |
|            System.Collections |          TryAddGiventSize&lt;Int32&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |        62,035.232 ns |       625.4011 ns |       554.4018 ns |        61,999.683 ns |        61,072.372 ns |        63,033.053 ns |      7.4111 |      3.7055 |           - |             46784 B |
|            System.Collections |         TryAddGiventSize&lt;String&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |       147,364.413 ns |     2,351.7824 ns |     2,084.7938 ns |       146,609.444 ns |       145,433.525 ns |       152,536.152 ns |     10.5140 |      5.2570 |           - |             65448 B |
|            System.Collections |            AddDefaultSize&lt;Int32&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |       151,457.558 ns |     2,898.9032 ns |     2,711.6358 ns |       150,197.879 ns |       148,626.916 ns |       156,803.193 ns |     24.0385 |     12.0192 |           - |            154192 B |
|            System.Collections |           AddDefaultSize&lt;String&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |       258,826.523 ns |     1,935.8032 ns |     1,810.7515 ns |       258,518.343 ns |       255,039.852 ns |       261,568.971 ns |     46.1066 |     44.0574 |     26.6393 |            215722 B |
|            **System.Collections** |                    **Remove&lt;Int32&gt;** |    **Dictionary** |            **1000** |            **1** | **2048** |       **?** |     **?** |        **62,786.129 ns** |       **834.9818 ns** |       **740.1896 ns** |        **62,748.654 ns** |        **61,766.815 ns** |        **64,588.044 ns** |           **-** |           **-** |           **-** |                   **-** |
|            System.Collections |                   Remove&lt;String&gt; |    Dictionary |            1000 |            1 | 2048 |       ? |     ? |       133,691.245 ns |     2,540.5734 ns |     2,376.4540 ns |       132,631.167 ns |       130,215.066 ns |       137,910.659 ns |           - |           - |           - |                   - |
|            System.Collections |                     Clear&lt;Int32&gt; |    Dictionary |            1000 |            1 | 2048 |       ? |     ? |         8,307.190 ns |       192.6903 ns |       214.1748 ns |         8,305.994 ns |         7,901.233 ns |         8,650.261 ns |           - |           - |           - |                   - |
|            System.Collections |                    Clear&lt;String&gt; |    Dictionary |            1000 |            1 | 2048 |       ? |     ? |        10,672.075 ns |       140.5965 ns |       117.4045 ns |        10,662.063 ns |        10,496.712 ns |        10,923.105 ns |           - |           - |           - |                   - |
|            System.Collections |   ContainsKeyFalse&lt;Int32, Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        46,885.587 ns |       552.1149 ns |       516.4486 ns |        46,941.644 ns |        45,685.077 ns |        47,682.228 ns |           - |           - |           - |                   - |
|            System.Collections | ContainsKeyFalse&lt;String, String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |       109,510.313 ns |       851.8671 ns |       796.8370 ns |       109,437.392 ns |       108,226.451 ns |       110,776.526 ns |           - |           - |           - |                   - |
|            System.Collections |    ContainsKeyTrue&lt;Int32, Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        42,012.134 ns |       350.9807 ns |       328.3075 ns |        42,010.042 ns |        41,423.956 ns |        42,699.365 ns |           - |           - |           - |                   - |
|            System.Collections |  ContainsKeyTrue&lt;String, String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |       109,269.743 ns |     1,116.3895 ns |     1,044.2715 ns |       109,322.904 ns |       108,074.619 ns |       111,986.671 ns |           - |           - |           - |                   - |
| System.Collections.Concurrent |                     Count&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        21,219.886 ns |       640.3695 ns |       737.4505 ns |        21,284.655 ns |        19,958.709 ns |        22,553.940 ns |           - |           - |           - |                   - |
| System.Collections.Concurrent |                    Count&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        20,212.199 ns |       175.6823 ns |       164.3333 ns |        20,235.303 ns |        19,916.917 ns |        20,478.918 ns |           - |           - |           - |                   - |
| System.Collections.Concurrent |                   IsEmpty&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |             2.983 ns |         0.1841 ns |         0.1722 ns |             2.970 ns |             2.709 ns |             3.323 ns |           - |           - |           - |                   - |
| System.Collections.Concurrent |                  IsEmpty&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |             3.017 ns |         0.1457 ns |         0.1291 ns |             2.981 ns |             2.854 ns |             3.276 ns |           - |           - |           - |                   - |
|            System.Collections |              AddGivenSize&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        63,489.339 ns |     1,488.1789 ns |     1,713.7890 ns |        63,791.385 ns |        60,512.448 ns |        66,618.257 ns |      7.4588 |      3.6008 |           - |             46784 B |
|            System.Collections |             AddGivenSize&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |       145,403.806 ns |     1,692.4725 ns |     1,583.1399 ns |       145,222.810 ns |       143,469.853 ns |       149,133.757 ns |     10.4167 |      5.2083 |           - |             65448 B |
|            System.Collections |        CtorFromCollection&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        64,083.569 ns |       707.7096 ns |       661.9920 ns |        64,151.889 ns |        63,065.820 ns |        65,028.320 ns |      7.4897 |      3.6157 |           - |             46784 B |
|            System.Collections |       CtorFromCollection&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |       150,999.660 ns |     2,020.8942 ns |     1,890.3456 ns |       150,848.377 ns |       147,695.729 ns |       155,176.823 ns |     10.4167 |      5.2083 |           - |             65448 B |
|            System.Collections |             CtorGivenSize&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        12,027.516 ns |       209.0856 ns |       195.5788 ns |        12,082.695 ns |        11,642.689 ns |        12,331.592 ns |      7.4472 |      3.7236 |           - |             46784 B |
|            System.Collections |            CtorGivenSize&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        23,727.781 ns |       369.4205 ns |       345.5561 ns |        23,688.431 ns |        23,077.184 ns |        24,386.122 ns |     10.3855 |      5.1460 |           - |             65448 B |
|            System.Collections |                IndexerSet&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        50,974.114 ns |       941.5698 ns |       880.7450 ns |        50,592.549 ns |        49,743.466 ns |        52,757.452 ns |           - |           - |           - |                   - |
|            System.Collections |               IndexerSet&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |       124,769.958 ns |     1,494.9546 ns |     1,398.3814 ns |       124,507.891 ns |       122,417.082 ns |       127,556.236 ns |           - |           - |           - |                   - |
|            System.Collections |            IterateForEach&lt;Int32&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        18,215.827 ns |       203.7896 ns |       190.6249 ns |        18,195.956 ns |        17,896.050 ns |        18,554.976 ns |           - |           - |           - |                   - |
|            System.Collections |           IterateForEach&lt;String&gt; |    Dictionary |               1 |           16 | 2048 |       ? |     ? |        18,497.016 ns |       190.6483 ns |       178.3325 ns |        18,461.196 ns |        18,305.749 ns |        18,943.396 ns |           - |           - |           - |                   - |
|      **System.Collections.Tests** |                  **Perf_Dictionary** | **ContainsValue** |               **1** |           **16** |    **?** |    **3000** |     **?** |     **7,479,647.883 ns** |    **84,846.0839 ns** |    **79,365.0796 ns** |     **7,503,163.724 ns** |     **7,335,345.834 ns** |     **7,575,527.598 ns** |           **-** |           **-** |           **-** |                   **-** |
|            **System.Collections** |        **DictionaryMappingFunction** |       **Entropy** |               **1** |           **16** |    **?** |    **5000** |     **?** |     **1,047,806.953 ns** |    **12,991.8550 ns** |    **12,152.5892 ns** |     **1,049,529.001 ns** |     **1,028,935.009 ns** |     **1,069,438.614 ns** |    **120.8333** |    **120.8333** |    **120.8333** |            **673056 B** |
|            **System.Collections** |        **DictionaryMappingFunction** |       **Entropy** |               **1** |           **16** |    **?** |   **50000** |     **?** |     **8,611,516.872 ns** |   **160,790.0849 ns** |   **157,917.3629 ns** |     **8,648,810.877 ns** |     **8,290,096.415 ns** |     **8,892,345.790 ns** |   **1062.5000** |   **1031.2500** |   **1031.2500** |           **6037793 B** |
|            **System.Collections** |        **DictionaryMappingFunction** |       **Entropy** |               **1** |           **16** |    **?** |  **500000** |     **?** |   **115,474,366.049 ns** | **2,256,936.6814 ns** | **2,216,613.6005 ns** |   **115,663,043.738 ns** |   **111,832,845.410 ns** |   **118,933,395.905 ns** |   **2500.0000** |   **2500.0000** |   **2500.0000** |          **53888568 B** |
|            **System.Collections** |        **DictionaryMappingFunction** |       **Entropy** |               **1** |           **16** |    **?** | **5000000** |     **?** | **1,060,218,842.288 ns** | **6,433,913.8710 ns** | **6,018,287.0268 ns** | **1,058,696,955.310 ns** | **1,048,741,079.400 ns** | **1,071,752,699.780 ns** |   **4000.0000** |   **4000.0000** |   **4000.0000** |         **471722104 B** |

After 

``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=Windows 10.0.17134.590 (1803/April2018Update/Redstone4)
Intel Core i7 CPU 860 2.80GHz (Nehalem), 1 CPU, 8 logical and 4 physical cores
Frequency=2727535 Hz, Resolution=366.6314 ns, Timer=TSC
.NET Core SDK=3.0.100-preview-010184
  [Host]     : .NET Core 3.0.0-preview-27324-5 (CoreCLR 4.6.27322.0, CoreFX 4.7.19.7311), 64bit RyuJIT
  Job-BXGUAA : .NET Core 87f9cb2e-5075-4cbd-ab6c-b5da46079c94 (CoreCLR 4.6.27523.0, CoreFX 4.7.19.12501), 64bit RyuJIT
  Job-KJQYMZ : .NET Core 87f9cb2e-5075-4cbd-ab6c-b5da46079c94 (CoreCLR 4.6.27523.0, CoreFX 4.7.19.12501), 64bit RyuJIT

Runtime=Core  Toolchain=CoreRun  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                     Namespace |                             Type |        Method | InvocationCount | UnrollFactor | Size |   Items | Count |                 Mean |             Error |            StdDev |               Median |                  Min |                  Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------ |--------------------------------- |-------------- |---------------- |------------- |----- |-------- |------ |---------------------:|------------------:|------------------:|---------------------:|---------------------:|---------------------:|------------:|------------:|------------:|--------------------:|
|            **System.Collections** |           **CtorDefaultSize&lt;Int32&gt;** |    **Dictionary** |               **1** |           **16** |    **?** |       **?** |     **?** |            **23.805 ns** |         **0.2385 ns** |         **0.2231 ns** |            **23.704 ns** |            **23.541 ns** |            **24.239 ns** |      **0.0171** |           **-** |           **-** |                **72 B** |
|            System.Collections |          CtorDefaultSize&lt;String&gt; |    Dictionary |               1 |           16 |    ? |       ? |     ? |           113.941 ns |         0.9883 ns |         0.9245 ns |           113.855 ns |           112.458 ns |           115.495 ns |      0.0170 |           - |           - |                72 B |
| **System.Collections.Concurrent** |                   **IsEmpty&lt;Int32&gt;** |    **Dictionary** |               **1** |           **16** |    **0** |       **?** |     **?** |           **170.869 ns** |         **1.9216 ns** |         **1.7975 ns** |           **170.418 ns** |           **166.628 ns** |           **173.358 ns** |           **-** |           **-** |           **-** |                   **-** |
| System.Collections.Concurrent |                  IsEmpty&lt;String&gt; |    Dictionary |               1 |           16 |    0 |       ? |     ? |           179.837 ns |         1.7561 ns |         1.6427 ns |           180.384 ns |           176.214 ns |           181.830 ns |           - |           - |           - |                   - |
|            **System.Collections** |        **DictionaryMappingFunction** |       **Entropy** |               **1** |           **16** |    **?** |     **500** |     **?** |        **82,340.128 ns** |       **978.2604 ns** |       **915.0654 ns** |        **82,026.180 ns** |        **81,309.353 ns** |        **83,891.009 ns** |     **17.3429** |      **0.3272** |           **-** |             **73160 B** |
|            **System.Collections** |         **TryAddDefaultSize&lt;Int32&gt;** |    **Dictionary** |               **1** |           **16** |    **?** |       **?** |  **2048** |       **150,153.371 ns** |     **1,618.9973 ns** |     **1,514.4110 ns** |       **150,124.634 ns** |       **147,893.717 ns** |       **152,240.763 ns** |     **24.1745** |     **11.7925** |           **-** |            **154192 B** |
|            System.Collections |        TryAddDefaultSize&lt;String&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |       261,100.120 ns |     1,934.8411 ns |     1,809.8516 ns |       261,193.832 ns |       257,340.070 ns |       264,332.737 ns |     46.1066 |     39.9590 |     26.6393 |            215717 B |
|            System.Collections |          TryAddGiventSize&lt;Int32&gt; |    Dictionary |               1 |           16 |    ? |       ? |  2048 |        63,706.533 ns |     2,213.4287 ns |     2,548.9878 ns |        63,218.560 ns |        60,307.853 ns |        67,743.168 ns |      7.2614 |      3.6307 |           - |             46784 B |
…