@MarcoRossignoli 2019-02-27
@dotnet-bot test this please