@MarcoRossignoli 2019-02-27
@safern should new command works also here?