@safern 2019-02-27
> @safern should new command works also here?

Yeah, it should work. Weird that it didn't work for you. Let me give it a try.