@danmosemsft 2019-03-02
@MarcoRossignoli thanks for looking at this. I am trying to absorb the information above. 

When gathering perf data, it is ideal to include the before and after in the same Benchmark.NET run so that it can show before and after in the same table, and calculate the base/candidate ratio (I see at least one of your tables has this) and also to be super explicit about what commit is base and what commit is candidate. That way it's easier for the rest of us to keep up 😃 

Of the various changes suggested in the issue, did any produce a clear perf improvement, and no significant regressions? Can we limit this PR to just that/those, and show the before/after perf data for just that/those?