@MarcoRossignoli 2019-03-03
> When gathering perf data, it is ideal to include the before and after in the same Benchmark.NET run so that it can show before and after in the same table, and calculate the base/candidate ratio (I see at least one of your tables has this) and also to be super explicit about what commit is base and what commit is candidate. That way it's easier for the rest of us to keep up 😃

I used new comparer to compare before/after, because AFAIK it's not possible merge 2 different benchmark run result togheter(@adamsitnik could you confirm?). Dictionary is on coreclr so to do tests I need to compile my local coreclr+corefx and after use performance repo tests on my local build. But I cannot run and compare old and new version togheter, so my strategy was run old code(coreclr+corefx with no dictionary update) and after new updated code(updated dic on coreclr+corefx) and compare with comparer tool. BTW now I'll try to do comparison on same report cloning 4 repo, current coreclr+corefx vs coreclr(dic updated)+corefx using two CoreRun.exe, I think it's the only way compare in a correct way what do you think @adamsitnik?Is there better way or a benchmarkdotnet feature to merge two "cold" result?

> Of the various changes suggested in the issue, did any produce a clear perf improvement, and no significant regressions? Can we limit this PR to just that/those, and show the before/after perf data for just that/those?

The issue asked to test 2 remaining point:
1) Use uint hash code for better entropy -> with uint hashcode we've slighly better performance in case of buckets collision, I added new tests to performance repo where I simulate int vs uint hash collision https://github.com/dotnet/performance/compare/master...MarcoRossignoli:newbenchdic

| Slower                                                              | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.CtorDefaultSize<Int32>.Dictionary                |      1.13 |            21.03 |            23.70 |         |
| System.Collections.Tests.Perf_Dictionary.ContainsValue(Items: 3000) |      1.06 |       7503163.72 |       7981264.83 |         |
| System.Collections.AddDefaultSize<String>.Dictionary(Count: 2048)   |      1.04 |        258518.34 |        268283.29 |         |

| Faster                                                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.DictionaryMappingFunction.Entropy(Items: 50000)       |      1.08 |       8648810.88 |       8006054.46 |         |
| System.Collections.AddGivenSize<Int32>.Dictionary(Size: 2048)            |      1.05 |         63791.39 |         60721.55 |         |
| System.Collections.ContainsKeyFalse<Int32, Int32>.Dictionary(Size: 2048) |      1.05 |         46941.64 |         44902.33 |         |
| System.Collections.Concurrent.Count<Int32>.Dictionary(Size: 2048)        |      1.04 |         21284.66 |         20467.29 |         |
| System.Collections.DictionaryMappingFunction.Entropy(Items: 500)         |      1.04 |         85026.70 |         82026.18 |         |
| System.Collections.Remove<String>.Dictionary(Size: 2048)                 |      1.03 |        132631.17 |        128484.00 |         |
| System.Collections.DictionaryMappingFunction.Entropy(Items: 5000000)     |      1.03 |    1058696955.31 |    1028937483.85 |         |
| System.Collections.TryAddGiventSize<String>.Dictionary(Count: 2048)      |      1.03 |        146609.44 |        142806.87 |         |

Check new `DictionaryMappingFunction`. Clearly the difference cannot be "so great" we are talking about few line of code(with few change in emitted code), improve entropy lead to "less" chain loop in case of collision, and this is more valuable in deep chain, my collision test has got max 2 element per collision, other perf remain basically unchanged(I did more than one test and every time slower results tests changed, the only thing that remain stable better/equal is new mapping function tests).
Collision test core code
```cs
for (int i = 1; i < Items; i++)
{
 dict.Add(i, i);
 dict.Add(int.MinValue + i, int.MinValue + i);
}
```

2) Eliminate the _freeCount field and use _freeList == -1 as a sentinel instead` -> As explained above I think that this change not worth it because if we remove _freecount we need to use _count as a real "non free list" item count, this lead to change every piece of code that "enumerate" the _entries(we need to use a local counter to decrement). After some preliminary tests I found that the code perform as expected worste, due to more complexity of scans,resize(trim/ensure capacity) and enumeration. This alg works better on DictionarySlim because it has less features for instance trim, ensure capacity, more enumerators type with different behaviours, contructors and features like remove item during enumeration.

Finally I think that uint as hashcode could be an improvement in case of collision with not measurable regression, eliminate _freeCount remove one local var but regress all other features due to more complexity on _entry enumeration(after this PR I'll show results on this second one if needed)

This PR show result only of one reasonable improvement "Use uint hash code for better entropy".

/cc @danmosemsft