@MarcoRossignoli 2019-03-03
@dotnet-bot test this please