@MarcoRossignoli 2019-03-04
rebased for better comparison with coreclr+corefx upstream