@danmosemsft 2019-03-04
Thanks @MarcoRossignoli that makes things clearer. For your perf results immediately above, is it possible to include the error? For example I am not sure how CtorDefaultSize could possibly change, but perhaps that 13% is within the error since it is a very short time. Likewise I wonder how the improvements compare to the noise level.